+++ application/controllers/Fleet.php
@@ -86,8 +86,9 @@ class Fleet extends Application
     private function showit()
     {
         $this->load->helper('form');
-        $plane = $this->session->userdata('plane');
-        $this->data['id'] = $plane->id;
+        $plane = (array)$this->session->userdata('plane');
+        $this->session->set_userdata('plane', $plane);
+        $this->data['id'] = $plane['id'];
 
         // if no errors, pass an empty message
         if (!isset($this->data['error'])) {
@@ -95,16 +96,16 @@ class Fleet extends Application
         }
 
         $fields = array(
-            'fid'      => form_label('ID') . form_input('plane', $plane->id),
-            'fmake'      => form_label('Make') . form_input('plane', $plane->manufacturer),
-            'fmodel'      => form_label('Model') . form_input('plane', $plane->model),
-            'fprice'      => form_label('Price') . form_input('plane', $plane->price),
-            'fseats'      => form_label('Seats') . form_input('plane', $plane->seats),
-            'freach'      => form_label('Reach') . form_input('plane', $plane->reach),
-            'fcruise'      => form_label('Cruise') . form_input('plane', $plane->cruise),
-            'ftakeoff'      => form_label('Takeoff') . form_input('plane', $plane->takeoff),
-            'fhourly'      => form_label('Hourly') . form_input('plane', $plane->hourly),
-            'zsubmit'    => form_submit('submit', 'Update Fleet'),
+            'fid'      => form_label('ID') . form_input('id', $plane['id']),
+            'fmake'      => form_label('Make') . form_input('make', $plane['manufacturer']),
+            'fmodel'      => form_label('Model') . form_input('model', $plane['model']),
+            'fprice'      => form_label('Price') . form_input('price', $plane['price']),
+            'fseats'      => form_label('Seats') . form_input('seats', $plane['seats']),
+            'freach'      => form_label('Reach') . form_input('reach', $plane['reach']),
+            'fcruise'      => form_label('Cruise') . form_input('cruise', $plane['cruise']),
+            'ftakeoff'      => form_label('Takeoff') . form_input('takeoff', $plane['takeoff']),
+            'fhourly'      => form_label('Hourly') . form_input('hourly', $plane['hourly']),
+            'zsubmit'    => form_submit('submit', 'Update Plane'),
         );
         $this->data = array_merge($this->data, $fields);
 
@@ -116,10 +117,10 @@ class Fleet extends Application
     public function submit()
     {
         // setup for validation
+        $this->load->helper('form');
         $this->load->library('form_validation');
         $this->form_validation->set_rules($this->fleetmdl->rules());
 
-        echo "EHITHITHTIHITIHT";
 
         // retrieve & update data transfer buffer
         $plane = (array) $this->session->userdata('plane');