+++ application/models/Fleetmdl.php
@@ -37,7 +37,7 @@ class Fleetmdl extends CSV_Model
           }
         }
 
-				return $remotePlane;
+        return $remotePlane;
     }
 
     // provide form validation rules